debug

+++ includes/finalize.sh
@@ -1,6 +1,8 @@
-if [ "$NGINX_ENABLED" = "true" ]; then
-    sudo systemctl stop nginx
-    sudo systemctl start nginx
+if [ "$ProxyEnabled_FLAG" = "true" ]; then
+    if [ "$NGINX_ENABLED" = "true" ]; then
+        sudo systemctl stop nginx
+        sudo systemctl start nginx
+    fi
 fi
 
 echo "Your application is now fully configured."
